cloud_functions/gps_publisher_fn /requirements.txt
pandas
functions-framework
google-cloud-storage
google-cloud-pubsub
flask